optimized dashboard code
