+++ web/AASTU-web-group-1/aastu.web.g1.bank-dashboard/components/ui/toast.tsx
@@ -32,9 +32,9 @@ const toastVariants = cva(
         default:
           "border bg-white text-slate-950 dark:bg-slate-950 dark:text-slate-50",
         destructive:
-          "destructive group border-red-500 bg-red-500 text-slate-50 dark:border-red-900 dark:bg-red-900 dark:text-slate-50",
+          "destructive group border-red-500 border-2 bg-white text-black dark:border-red-900 dark:bg-red-900 dark:text-slate-50",
         success:
-          "success group border-green-500 bg-green-500 text-slate-50 dark:border-green-900 dark:bg-green-900 dark:text-slate-50",
+          "success group border-green-500 border-2 bg-white text-black dark:border-green-900 dark:bg-green-900 dark:text-slate-50",
       },
     },
     defaultVariants: {
